+++ fabric/src/main/java/io/github/sakurawald/fuji/core/event/message/player/PlayerChatMessageSentEvent.java
@@ -6,12 +6,14 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import net.minecraft.network.message.MessageType;
 import net.minecraft.network.message.SignedMessage;
+import net.minecraft.server.network.ServerPlayerEntity;
 
 @Data
 @AllArgsConstructor
 @EqualsAndHashCode(callSuper = true)
 public class PlayerChatMessageSentEvent extends BaseEvent {
 
+    ServerPlayerEntity receiverPlayer;
     SignedMessage signedMessage;
     MessageType.Parameters parameters;