+++ fabric/src/main/java/io/github/sakurawald/fuji/module/mixin/core/event/PlayerChatMessageSentEventMixin.java
@@ -3,15 +3,19 @@ package io.github.sakurawald.fuji.module.mixin.core.event;
 import com.google.common.collect.EvictingQueue;
 import io.github.sakurawald.annotation.PhasedMixinTemplate;
 import io.github.sakurawald.auxiliary.WeaverUtil;
+import io.github.sakurawald.fuji.core.auxiliary.minecraft.RegistryHelper;
 import io.github.sakurawald.fuji.core.event.EventManager;
 import io.github.sakurawald.fuji.core.event.annotation.EventProducer;
 import io.github.sakurawald.fuji.core.event.message.player.PlayerChatMessageSentEvent;
+import java.util.Objects;
 import java.util.Queue;
 import net.minecraft.network.message.MessageType;
 import net.minecraft.network.message.SignedMessage;
 import net.minecraft.server.network.ServerPlayNetworkHandler;
+import net.minecraft.server.network.ServerPlayerEntity;
 import org.jetbrains.annotations.NotNull;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -21,15 +25,20 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(value = ServerPlayNetworkHandler.class)
 public abstract class PlayerChatMessageSentEventMixin {
 
+    @Shadow
+    public abstract ServerPlayerEntity getPlayer();
+
     @Unique
     private static final Queue<Long> DUPLICATED_SENT_TEXT_FILTER = EvictingQueue.create(8);
 
     @Unique
-    private static long toUniqueKey(@NotNull SignedMessage signedMessage) {
+    private static long toUniqueKey(@NotNull SignedMessage signedMessage, MessageType.Parameters parameters) {
         // NOTE: The SignedMessage#getSalt method only works in online-mode server. In offline-mode server, it always returns 0.
         // NOTE: The hashCode() is used as the distinguish key, because the SentMessage#send is called inside a loop, and will not be modified.
         // NOTE: For chat-related mod compatibility, here I have to capture the chat message from network layer, and do the filter things. I want to ensure that's the last possible point the chat message can be processed.
-        return signedMessage.hashCode();
+        String messageTypeString = RegistryHelper.getIdAsString(parameters);
+
+        return Objects.hash(signedMessage.hashCode(), messageTypeString.hashCode());
     }
 
     /**
@@ -40,14 +49,15 @@ public abstract class PlayerChatMessageSentEventMixin {
     @Inject(method = "sendChatMessage", at = @At(value = "RETURN"))
     void producePlayerChatMessageSentEvent(SignedMessage signedMessage, MessageType.Parameters parameters, CallbackInfo ci) {
         /* Filter duplicated messages. */
-        long uniqueKey = toUniqueKey(signedMessage);
+        long uniqueKey = toUniqueKey(signedMessage, parameters);
         if (DUPLICATED_SENT_TEXT_FILTER.contains(uniqueKey)) {
             return;
         }
         DUPLICATED_SENT_TEXT_FILTER.add(uniqueKey);
 
         /* Produce the event. */
-        PlayerChatMessageSentEvent event = new PlayerChatMessageSentEvent(signedMessage, parameters);
+        ServerPlayerEntity receiverPlayer = getPlayer();
+        PlayerChatMessageSentEvent event = new PlayerChatMessageSentEvent(receiverPlayer, signedMessage, parameters);
         EventManager.dispatchEvent(PlayerChatMessageSentEvent.class, event, WeaverUtil.TOKEN_PLACEHOLDER);
     }
 }