+++ CHANGELOG.md
@@ -9,3 +9,7 @@
   - feature: fine-tune the event injection point, for better compatibility with other mods.
 - [command_spy] Enhance this module.
   - feature: fine-tune the event injection point, for better compatibility with other mods.
+
+- [chat] Enhance chat modules.
+  - [chat.history] feature: fine-tune the event injection point, for better compatibility with other mods.
+  - [chat.spy] feature: fine-tune the event injection point, for better compatibility with other mods.